Phase 3 Step 1-2: Extract exceptions.py and models.py
- Created src/batch/exceptions.py (~40 lines)
  - BatchEngineError, ConfigurationError, FileProcessingError, SecurityError

- Created src/batch/models.py (~84 lines)
  - BatchJob and BatchManifest dataclasses
  - BATCH_MANIFEST_FILENAME and JOBS_DIRNAME constants

- Updated src/batch/engine.py
  - Removed exception class definitions (~40 lines)
  - Removed dataclass definitions (~70 lines)
  - Added imports from exceptions and models modules
  - Line count reduced from 2,112 to ~2,002 lines

- Updated src/batch/__init__.py
  - Export exceptions and models for backward compatibility

✅ All 81 Phase 2 tests passing
✅ 70% coverage maintained on engine.py
✅ Refactoring: Issue #328
